refactor: export Storage in a parent class
