+++ src/RewardsDistributor/index.ts
@@ -1,0 +1,1 @@
+export { default as MorphoCompoundRewardsDistributor } from "./MorphoCompoundRewardsDistributor";
