+++ test/rewardsDistributor.test.ts
@@ -1,16 +1,17 @@
-import { MorphoCompoundDistributor } from "../src/RewardsDistributor";
-import { BigNumber, providers } from "ethers";
+import { MorphoCompoundRewardsDistributor } from "../src/RewardsDistributor";
+import { BigNumber, constants, providers } from "ethers";
 import { MorphoCompound__factory } from "@morpho-labs/morpho-ethers-contract";
 import addresses from "@morpho-labs/morpho-ethers-contract/lib/addresses";
-import { ages, allEpochs } from "../lib";
+import { ages, allEpochs } from "../src";
 import * as fs from "fs/promises";
 import { fetchUsers } from "../src/utils";
+import { expectBNApproxEquals } from "./ageOne/epochOne.test";
 jest.setTimeout(3_600_000);
 
 let desc = describe;
 if (process.env.SKIP_DISTRIBUTOR) desc = desc.skip;
-describe("MorphoRewardsDistributor for Compound only", () => {
-  let rewardsDistributor: MorphoCompoundDistributor;
+desc("MorphoRewardsDistributor for Compound only", () => {
+  let rewardsDistributor: MorphoCompoundRewardsDistributor;
   const provider = new providers.JsonRpcProvider(process.env.RPC_URL);
 
   beforeAll(async () => {
@@ -18,10 +19,10 @@ describe("MorphoRewardsDistributor for Compound only", () => {
     if (process.env.INIT_DUMP_BLOCK) {
       // Use a dump for initialization
       const dump = require(`../storages/dump-${process.env.INIT_DUMP_BLOCK}.json`);
-      rewardsDistributor = MorphoCompoundDistributor.initFromDump(provider, morphoCompound, dump);
+      rewardsDistributor = MorphoCompoundRewardsDistributor.initFromDump(provider, morphoCompound, dump);
       console.log("Morpho rewards distributor initialized from dump");
     } else {
-      rewardsDistributor = await MorphoCompoundDistributor.init(provider, morphoCompound);
+      rewardsDistributor = await MorphoCompoundRewardsDistributor.init(provider, morphoCompound);
       console.log("Morpho rewards distributor initialized");
       await saveStorage(rewardsDistributor.store);
     }
@@ -48,25 +49,42 @@ describe("MorphoRewardsDistributor for Compound only", () => {
     });
   });
 
-  it("should compute correct storage distribution for epoch 1 compared to the subgraph", async () => {
+  describe("Epoch one", () => {
     const epochOne = ages[0].epochs[0];
-    await rewardsDistributor.applyEpoch(epochOne);
-    const storage = rewardsDistributor.store;
-    // first retrieve the state of the subgraph at the end of the first epoch
-    const graphDistribution = await fetchUsers(ages[0].subgraphUrl, epochOne.finalBlock);
-    // then, compare if each user balance is correct
-    graphDistribution.map((user) =>
-      user.balances.map((balance) => {
-        const storageBalance = storage.users[user.address][balance.market.address];
-        expect(storageBalance).not.toBeUndefined();
-        expect(storageBalance.accruedMorpho).toEqual(balance.accumulatedMorpho.toString());
-        expect(storageBalance.userBorrowIndex).toEqual(balance.userBorrowIndex.toString());
-        expect(storageBalance.userSupplyIndex).toEqual(balance.userSupplyIndex.toString());
-        expect(storageBalance.userSupplyBalance).toEqual(balance.underlyingSupplyBalance.toString());
-        expect(storageBalance.userBorrowBalance).toEqual(balance.underlyingBorrowBalance.toString());
-      })
-    );
-    await saveStorage(storage);
+    const epochOneRoot = require("../distribution/proofs/proofs-1.json").root;
+    beforeAll(async () => {
+      await rewardsDistributor.applyEpoch(epochOne);
+    });
+
+    it("Should compute correct storage distribution for epoch 1 compared to the subgraph", async () => {
+      const storage = rewardsDistributor.store;
+      // first retrieve the state of the subgraph at the end of the first epoch
+      const graphDistribution = await fetchUsers(ages[0].subgraphUrl, epochOne.finalBlock);
+      // then, compare if each user balance is correct
+      expect(graphDistribution).toHaveLength(Object.keys(storage.users).length);
+      graphDistribution.map((user) =>
+        user.balances.map((balance) => {
+          const storageBalance = storage.users[user.address][balance.market.address];
+          expect(storageBalance).not.toBeUndefined();
+          expect(storageBalance.accruedMorpho).toEqual(balance.accumulatedMorpho.toString());
+          expect(storageBalance.userBorrowIndex).toEqual(balance.userBorrowIndex.toString());
+          expect(storageBalance.userSupplyIndex).toEqual(balance.userSupplyIndex.toString());
+          expect(storageBalance.userSupplyBalance).toEqual(balance.underlyingSupplyBalance.toString());
+          expect(storageBalance.userBorrowBalance).toEqual(balance.underlyingBorrowBalance.toString());
+        })
+      );
+      await saveStorage(storage);
+    });
+
+    it("Should distribute the correct amount of rewards", async () => {
+      const totalEmitted = await rewardsDistributor.totalEmitted();
+      console.log("totalEmitted", totalEmitted.toString(), "over", epochOne.totalEmission.toString());
+      expectBNApproxEquals(totalEmitted, epochOne.totalEmission.mul(constants.WeiPerEther), 1e9);
+    });
+    it("Should compute the correct merkle root", async () => {
+      const { root } = await rewardsDistributor.computeMerkleTree();
+      expect(root).toEqual(epochOneRoot);
+    });
   });
 });
